cs50w_finalproject/requirements.txt
Django
django-quill-editor